requirements.txt
boto3>=1.26
pandas>=1.5
pyarrow>=11.0
oracledb>=1.0
psycopg2-binary>=2.9
